refactor(store): remove lifecycle writer coordination
