+++ backend/component/review/approval.go
@@ -152,9 +152,7 @@ func (a *ApprovalEvaluator) ApplyApprovalTemplate(ctx context.Context, input App
 			return nil, workflowWrap(ErrorInternal, err, "failed to acquire Plan review lock for approval finding")
 		}
 	}
-	// The existing Issue row is the project-lifecycle fence for this update:
-	// project purge deletes Issues before the Project. Approval may finish while
-	// a soft-deleted Project still exists, but cannot write after purge passes it.
+	// Lock the Issue before comparing and updating its approval payload.
 	lockedIssue, err := lockIssue(ctx, tx, input.ProjectID, input.IssueUID)
 	if err != nil {
 		return nil, err
