+++ docs/design/saved-queries-model.md
@@ -749,73 +749,39 @@ saved_query)` — still group references — indexed
 per-principal *ordered* seeks (`O(K · page)`) and isolating the ACL index
 from content churn. A pure storage change behind the same API.
 
-**Lifecycle vs. project deletion.** `saved_query.project` is an FK with no
-cascade, so the two directions are fenced explicitly (AGENTS.md transaction
-lock-ordering):
+**Writer behavior.** `saved_query.project` is an FK with no cascade.
 
-- **Create requires an *active* project.** `CreateSavedQuery` does not go
+- **Create requires an active project.** `CreateSavedQuery` does not go
   through `nextProjectID` (IDs are `gen_random_uuid()`, not a project-scoped
-  sequence), so it must itself, in the insert transaction, lock the project
-  row and reject the write unless the project is active — a create racing a
-  purge fails cleanly (`FAILED_PRECONDITION`), never as an FK violation and
-  never leaving a row in a project mid-deletion.
-- **Star writes are child-before-parent; the parent fence covers only the
-  first star.** The lock rule treats an upsert as an existing-row lock, so
-  toggling or removing an **existing** `saved_query_star` row locks that
-  child row directly — no parent lock. Only the **first** star for a
-  (query, caller) inserts a child that cannot be locked in advance; that
-  case alone takes the parent fence (lock the `saved_query` row, reject as
-  `NotFound` if gone), the same missing-child carve-out `CreateSavedQuery`
-  uses for `project`. Its inserted key is novel, so it never contends with
-  purge's existing-child locks — no cross-order deadlock.
+  sequence), so it verifies the project before inserting.
+- **Star writes are child-before-parent.** The lock rule treats an upsert as
+  an existing-row lock, so toggling or removing an **existing**
+  `saved_query_star` row locks that child row directly. A first star verifies
+  its parent saved query before inserting the new child.
 - **Batch folder moves lock rows in primary-key order.**
   `MoveMySavedQueries` only *updates* existing `saved_query` rows (no
   new child), but "existing" is not enough on its own: two overlapping
   batches could grab their target rows in different scan orders and
   deadlock. So it locks its selected rows in full primary-key order
-  (`resource_id`) — the AGENTS.md batch rule — and a folder move touching a
-  row mid-purge simply updates zero rows.
+  (`resource_id`) — the AGENTS.md batch rule.
 - **Delete is child-before-parent, explicitly.** `DeleteSavedQuery` removes
   the query's `saved_query_star` rows before the `saved_query` row itself —
-  **not** via the FK cascade, which would lock the parent first — matching
-  the star and purge order so a delete racing a star toggle or a purge
-  cannot deadlock.
-- **Purge is child-before-parent, and re-parents human rows.** The
-  hard-delete path deletes the stars and saved queries of project service
-  accounts and workload identities (stars first), then reassigns the
-  remaining, human-created saved queries to the default project — locking
-  existing child rows ahead of their parents, per the rule. Because a
-  surviving row changes project mid-purge, **every non-purge writer scopes
-  its predicate by `(resource_id, project)`** — patch, delete, and the
-  first-star parent fence — so a write authorized in the purged project
-  cannot land on the reassigned row; it updates zero rows and surfaces as
-  NotFound (the fenced delete rolls back its star cleanup). The declared
-  lifecycle policy: writers on existing rows require the row in the
-  authorized project, not an active project — archived projects are already
-  unreachable through every read path, so a write racing an archive is an
-  ordinary serialization of concurrent requests, and restore does not
-  promise to resurrect a row whose authorized delete won that race. Only
-  creation requires an active project.
+  **not** via the FK cascade, which would lock the parent first — so a delete
+  racing a star toggle cannot deadlock.
+
+Project and instance archive and physical purge are best-effort lifecycle
+operations. A rare concurrent writer may fail with a PostgreSQL conflict or
+commit based on state it observed before the lifecycle transition. These
+outcomes are accepted and the user can retry the lifecycle operation or write.
 
 Two invariants cover every writer. **(1) Child before parent** for
-existing-row writes/deletes; the *only* parent-first step is a new-child
-*insert* (create → lock `project`; first star → lock `saved_query`), safe
-because its key cannot be locked in advance — the AGENTS.md missing-child
-carve-out. **(2) Any multi-row lock/update/delete acquires its rows in full
-primary-key order** — `saved_query` by `resource_id`, `saved_query_star` by
-`(saved_query, principal)` — so two operations over an overlapping set
-(delete vs. purge deleting the same query's stars; two batch folder moves;
-purge's own `saved_query` batch) can never lock in opposing orders. A plain
-`DELETE … WHERE` does not guarantee that order, so these paths take their
-row locks through an ordered `… ORDER BY <pk> FOR UPDATE` (or an ordered
-delete) before mutating. Required before
-implementation: deterministic real-PostgreSQL regression tests for **both**
-acquisition orders of each contending pair — create↔purge,
-first-star↔purge, delete↔star-toggle, delete↔purge over a query with
-multiple stars, and two overlapping `MoveMySavedQueries` over
-the same folder — asserting the terminal outcomes — project (or query) deleted, no orphaned
-saved query or star, and **no** FK failure or deadlock (`40P01`) in either
-direction (absence of `40P01` alone is insufficient).
+existing-row writes and deletes. **(2) Any multi-row lock, update, or delete
+acquires its rows in full primary-key order** — `saved_query` by
+`resource_id`, `saved_query_star` by `(saved_query, principal)` — so
+overlapping ordinary operations, such as a delete and star toggle or two
+folder moves, cannot lock in opposing orders. A plain `DELETE … WHERE` does
+not guarantee that order, so these paths take their row locks through an
+ordered `… ORDER BY <pk> FOR UPDATE` (or an ordered delete) before mutating.
 
 ### Sharing and organization UX
